implementing builders to allow for more customization
