+++ app/BuilderInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App;
+
+interface BuilderInterface
+{
+    public function buildSections(array $nodes): string;
+
+}
